DetectionFastAPI/requirements.txt
grpcio
grpcio-tools
numpy
opencv-python
ultralytics
torch
fastapi
uvicorn